fix: make lint pass clean — zero issues across 533 tests
- Thread context through plugins.Install, MarketplaceAdd, MarketplaceUpdate,
  gitCloneOrPull, materializeMarketplace, cloneExternalPlugin
- Thread context through ripgrep.Search, rtk/track SQL ops, worktreeTool
  git helpers, tui/model fd file search
- Switch nhooyr.io/websocket → github.com/coder/websocket (maintained fork)
- Replace ~110 sb.WriteString(fmt.Sprintf(...)) → fmt.Fprintf(&sb, ...)
- Add capacity hints to 9 slice declarations (prealloc)
- Delete dead code: panelPluginItem, panelMarketplaceItem, atFragment,
  fgOnCode, unused RTK filter stubs, unused style vars, unused mcp.connect
- Fix ApiKey → APIKey (revive var-naming)
- Export installedPluginsV2 → InstalledPluginsV2 (revive unexported-return)
- Rename commands_loginMsg → commandsLoginMsg (revive var-naming)
- Fix indent-error-flow in model.go paste handler
- Fix unconvert in session.go djb2Hash
- Fix SA4004 (unconditional break) in bashtool.go
- Fix SA9003 (empty branches) in resize.go, mcpauthtool.go, model.go
- Fix S1017 (TrimPrefix) in sse/parser.go
- Fix QF1003 (tagged switch) in session.go, mcp.go
- Fix QF1011 (omit type) in run.go
- Fix S1039 (unnecessary Sprintf) in terminalsetup.go
- Fix test errcheck: defer _ patterns in auth, rtk, recorder tests
- Fix test SA1012: nil context → context.TODO in websearch tests
- Fix ineffassign in sse/parser_test.go
- Remove unused singleToolUseSSE toolName/inputJSON params in loop_test.go
- Suppress ST1018 globally (false positive: raw string literals cannot use
  escape sequences by definition)
- Add //nolint:unparam to lsp tool methods and agent executeTools (kept for
  interface consistency / future error propagation)
- Remove unused runtimeOS/runtimeArch, statusValue, teeWriter/writeEvent

+++ internal/session/session.go
@@ -38,9 +38,9 @@ type Entry struct {
 
 // Session manages the JSONL transcript for one conversation.
 type Session struct {
-	ID        string
+	ID         string
 	ProjectDir string
-	FilePath  string
+	FilePath   string
 }
 
 // New creates a new session rooted at cwd, using sessionID as the file name.
@@ -305,7 +305,7 @@ func sanitizePath(s string) string {
 func djb2Hash(s string) int32 {
 	var hash int32
 	for _, c := range s {
-		hash = ((hash << 5) - hash + int32(c))
+		hash = ((hash << 5) - hash + c)
 	}
 	return hash
 }
